chore: reorganize organization folders for better coherence
