+++ client/src/app/components/manage-organizations/organization-list/manage-organizations.component.html
