+++ client/src/app/components/manage-organizations/organization-list/manage-organizations.component.scss
@@ -1,4 +1,4 @@
-@import "./src/theme.scss";
+@import "../../../../theme";
 
 .search-header {
   min-height: 64px;
