+++ client/src/app/components/manage-organizations/organization-list/organization-dialog.html
