+++ client/src/app/components/manage-organizations/organization-view/organization-view.component.scss
@@ -1,4 +1,4 @@
-@import "./src/theme.scss";
+@import "../../../../theme";
 
 .title-padding {
   padding: 1%;
