+++ client/src/app/components/manage-organizations/organization-view/organization-view.component.ts
